Updates July 2019
Includes the following fixes and updates:

## Release-then-rescind bug fixed

There was an issue where feedback being released, then *all* feedback for that user being rescinded, would result in the released feedback not being removed from the gradebook. This would only occur if there was no feedback from the plugin itself.

This commit fixes this issue.

## Fixed question duplication bug

Split100 and Comment type questions could become duplicated when reordering in some circumstances. This update fixes this issue.

## Fix report filenames

Filenames in teamevalreport_responses must replace slashes in the title with a non-path sensitive character. moodle_url escapes URL-sensitive characters... *except slashes*. It assumes that any slash character is part of the path. For this reason, we cannot url-encode the slashes before sending them to moodle_url, as it will result in double-encoding. Instead, we simply replace the slash with a colon, which will at least appear as a slash in the filename on macOS...

## Corrected "Public" help text

Public no longer works this way

---

These updates are the result of ongoing development at UNSW.